refactor: 검증 코드 개선
- 검증 관련 코드 메서드를 분리했다.
- 누락된 테스트 코드를 작성했다.
